Combine test scripts
