--- tests/test_proteus.py
@@ -1,21 +1,0 @@
-from __future__ import annotations
-
-from itertools import chain
-
-import pytest
-from helpers import PROTEUS_ROOT
-
-from proteus import Proteus
-from proteus.config import Config
-
-PATHS = chain(
-    (PROTEUS_ROOT / 'input').glob('*.toml'),
-    (PROTEUS_ROOT / 'examples').glob('*/*.toml'),
-)
-
-
-@pytest.mark.parametrize('path', PATHS)
-def test_proteus_init(path):
-    runner = Proteus(config_path=path)
-
-    assert isinstance(runner.config, Config)
